move the swaggergen logic to pkg/swaggergen
